+++ src/app/cart/page.tsx
@@ -1,15 +1,12 @@
 'use client';
 
 import { ItemCardList } from '@/components/cart/ItemList';
-import { QueryProvider } from '../provider';
 
 export default function CartPage() {
   return (
-    <QueryProvider>
-      <div className="my-10 flex flex-col items-center gap-5 p-10">
-        <h2 className="text-2xl font-bold">장바구니</h2>
-        <ItemCardList />
-      </div>
-    </QueryProvider>
+    <div className="my-10 flex flex-col items-center gap-5 p-10">
+      <h2 className="text-2xl font-bold">장바구니</h2>
+      <ItemCardList />
+    </div>
   );
 }
