+++ src/app/detail/[item_id]/page.tsx
@@ -5,9 +5,9 @@ import { useQuery } from '@tanstack/react-query';
 import supabase from '@/services/supabase';
 import { ImageSlider } from '@/components/detail/ImageSlider';
 import { AddCartButton } from '@/components/cart/AddCartButton';
-import {Loading} from '@/components/detail/Loading'; // 로딩 컴포넌트
-import {Error2} from '@/components/detail/Error'; 
-import {Item} from "@/types/type"
+import { Loading } from '@/components/detail/Loading'; // 로딩 컴포넌트
+import { Error2 } from '@/components/detail/Error';
+import { Item } from '@/types/cartType';
 
 // 특정 상품을 Supabase에서 가져오는 함수
 const fetchDetailItem = async (item_id: string | string[]): Promise<Item> => {
@@ -31,7 +31,7 @@ export default function DetailPage() {
     error,
   } = useQuery<Item>({
     queryKey: ['item', item_id], // item_id 별로 캐싱하는 쿼리키
-    queryFn: () => fetchDetailItem(item_id), 
+    queryFn: () => fetchDetailItem(item_id),
     enabled: !!item_id, // 상품 id 없을땐 쿼리 실행 안함
   });
 
@@ -52,10 +52,9 @@ export default function DetailPage() {
   // 에러 상태 처리
   if (isError) return <Error2 message={(error as Error).message} />;
 
-
   return (
     <main className="mx-auto max-w-6xl p-4 pt-20">
-       {/* 반응형 디자인: 모바일에서는 세로, 데스크탑에서는 가로 */}
+      {/* 반응형 디자인: 모바일에서는 세로, 데스크탑에서는 가로 */}
       <div className="flex flex-col gap-8 md:flex-row">
         {/* 왼쪽: 이미지 섹션 */}
         <div className="md:w-1/2">
@@ -63,13 +62,11 @@ export default function DetailPage() {
         </div>
 
         {/* 오른쪽: 상품 정보 섹션 */}
-        <div className="md:w-1/2 flex flex-col justify-between">
+        <div className="flex flex-col justify-between md:w-1/2">
           <div>
-            <h1 className="text-2xl font-bold mb-4">{item!.title}</h1>
-            <p className="text-base text-gray-700 mb-2">{item!.content}</p>
-            <p className="text-lg font-semibold mb-6 mt-3">
-              {item!.price.toLocaleString()}원
-            </p>
+            <h1 className="mb-4 text-2xl font-bold">{item!.title}</h1>
+            <p className="mb-2 text-base text-gray-700">{item!.content}</p>
+            <p className="mb-6 mt-3 text-lg font-semibold">{item!.price.toLocaleString()}원</p>
           </div>
 
           {/* 장바구니 버튼 추가 */}
