+++ src/app/page.tsx
@@ -4,10 +4,11 @@ import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import Image from 'next/image';
 import supabase from '@/services/supabase';
-import { Item } from '@/types/type';
+
 import { Loading } from '@/components/detail/Loading';
 import { Error2 } from '@/components/detail/Error';
 import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
+import { Item } from '@/types/cartType';
 
 // Supabase에서 전체 상품 리스트를 불러오는 함수
 async function fetchItems(): Promise<Item[]> {
